1. Removed the overridden "initiateAuthenticationRequest" method and use the super method instead. 2. Removed the overridden "processAuthenticationResponse" method and use the super method instead. 3. Moved "sendRequest" method to super. 4. Stopped overriding the "getSubjectAttributes" method and created a new override method to get the user claims URL. 5. Getting the default URLs from a configuration file.
