+++ components/application-authenticators/org.wso2.carbon.identity.application.authenticator.oidc/src/main/java/org/wso2/carbon/identity/application/authenticator/oidc/OpenIDConnectAuthenticator.java
@@ -17,6 +17,7 @@
  */
 package org.wso2.carbon.identity.application.authenticator.oidc;
 
+import org.apache.commons.lang.StringUtils;
 import org.apache.oltu.oauth2.client.OAuthClient;
 import org.apache.oltu.oauth2.client.URLConnectionClient;
 import org.apache.oltu.oauth2.client.request.OAuthClientRequest;
@@ -42,7 +43,11 @@ import org.wso2.carbon.identity.core.util.IdentityUtil;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.InputStreamReader;
+import java.net.HttpURLConnection;
+import java.net.URL;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -127,11 +132,62 @@ public class OpenIDConnectAuthenticator extends AbstractApplicationAuthenticator
     }
 
     /**
+     *
+     * @param token
+     * @return
+     */
+    protected String getUserClaimUri(OAuthClientResponse token) {
+        return null;
+    }
+
+    /**
      * @param token
      * @return
      */
     protected Map<ClaimMapping, String> getSubjectAttributes(OAuthClientResponse token) {
-        return new HashMap<ClaimMapping, String>();
+
+        Map<ClaimMapping, String> claims = new HashMap<ClaimMapping, String>();
+
+        try {
+
+            String accessToken = token.getParam(OIDCAuthenticatorConstants.ACCESS_TOKEN);
+            String url = getUserClaimUri(token);
+
+            String json = sendRequest(url, accessToken);
+
+            if (!StringUtils.isNotBlank(json)) {
+                log.info("Unable to fetch user claims. Proceeding without user claims");
+                return claims;
+            }
+
+            Map<String, Object> jsonObject = JSONUtils.parseJSON(json);
+
+            // Extract the inside profile JSON object.
+            Map<String, Object> profile = JSONUtils.parseJSON(
+                    jsonObject.entrySet().iterator().next().getValue().toString());
+
+            if (profile == null) {
+                log.info("Invalid user profile object. Proceeding without user claims");
+                return claims;
+            }
+
+            for (Map.Entry<String, Object> data : profile.entrySet()) {
+
+                String key = data.getKey();
+
+                claims.put(ClaimMapping.build(key, key, null, false), profile.get(key).toString());
+
+                if (log.isDebugEnabled()) {
+                    log.debug("Adding claims from end-point data mapping : " + key + " - " +
+                            profile.get(key).toString());
+                }
+            }
+
+        } catch (Exception e) {
+            log.error("Error occurred while accessing yahoo user info endpoint", e);
+        }
+
+        return claims;
     }
 
     @Override
@@ -450,5 +506,48 @@ public class OpenIDConnectAuthenticator extends AbstractApplicationAuthenticator
             }
         }
         return subject;
+    }
+
+    /**
+     * Request user claims from user info endpoint.
+     * @param url User info endpoint.
+     * @param accessToken Access token.
+     * @return Response string.
+     * @throws IOException
+     */
+    private String sendRequest(String url, String accessToken)
+            throws IOException {
+
+        if (log.isDebugEnabled()) {
+            log.debug("Claim URL: " + url + " & Access-Token : " + accessToken);
+        }
+
+        if (url == null) {
+            return StringUtils.EMPTY;
+        }
+
+        URL obj = new URL(url);
+        HttpURLConnection urlConnection = (HttpURLConnection) obj.openConnection();
+
+        urlConnection.setRequestMethod("GET");
+        urlConnection.setRequestProperty("Authorization", "Bearer " + accessToken);
+
+        BufferedReader reader = new BufferedReader(new InputStreamReader(urlConnection.getInputStream()));
+        StringBuilder builder = new StringBuilder();
+
+        String inputLine = reader.readLine();
+
+        while (inputLine != null) {
+            builder.append(inputLine).append("\n");
+            inputLine = reader.readLine();
+        }
+
+        reader.close();
+
+        if (log.isDebugEnabled()) {
+            log.debug("response: " + builder.toString());
+        }
+
+        return builder.toString();
     }
 }
