+++ features/authentication-framework/org.wso2.carbon.identity.application.authentication.framework.server.feature/resources/application-authentication.xml
@@ -89,7 +89,12 @@
 			<Parameter name="TrustStorePassword">wso2carbon</Parameter>
 			<!--Parameter name="OpenIDManager">org.wso2.carbon.identity.application.authenticator.openid.manager.DefaultOpenIDManager</Parameter>
 			<Parameter name="AttributesRequestor">org.wso2.carbon.identity.application.authenticator.openid.manager.SampleAttributesRequestor</Parameter-->
-		</AuthenticatorConfig>	
+		</AuthenticatorConfig>
+		<AuthenticatorConfig name="YahooOAUth2OpenIDAuthenticator" enabled="true">
+			<Parameter name="YahooTokenEndpoint">https://api.login.yahoo.com/oauth2/get_token</Parameter>
+			<Parameter name="YahooOAuth2Endpoint">https://api.login.yahoo.com/oauth2/request_auth</Parameter>
+			<Parameter name="YahooUserInfoURL">https://social.yahooapis.com/v1/user/</Parameter>
+		</AuthenticatorConfig>
 	</AuthenticatorConfigs> 
 
     <!--
